Ajustes de mensagens de validacao

+++ src/use_cases/customer/deleteById.js
@@ -1,7 +1,9 @@
 export default function deleteById(id, custumerRepository) {
     return custumerRepository.findById(id).then((custumer) => {
-      if (!custumer) {
-        throw new Error(`No custumer found with id: ${id}`);
+      //console.log('String(custumer)')
+      if (!custumer || String(custumer) === 'null') {
+        //throw new Error(`No custumer found with id: ${id}`);
+        return Promise.resolve(`No custumer found with id: ${id}`);
       }
       return custumerRepository.deleteById(id);
     });
